+++ app/src/main/java/com/denihilhamsyah/totphub/totp/presentation/TOTPEvent.kt
@@ -1,5 +1,7 @@
 package com.denihilhamsyah.totphub.totp.presentation
 
+import com.denihilhamsyah.totphub.totp.presentation.component.UiText
+
 sealed interface TOTPEvent {
     data class OnSecretAdded(val uiText: UiText) : TOTPEvent
     data class OnOperationFailed(val uiText: UiText) : TOTPEvent
