+++ app/src/main/java/com/denihilhamsyah/totphub/totp/presentation/component/ThemeSwitch.kt
@@ -1,4 +1,4 @@
-package com.denihilhamsyah.totphub.totp.presentation
+package com.denihilhamsyah.totphub.totp.presentation.component
 
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
