+++ app/src/main/java/com/denihilhamsyah/totphub/totp/presentation/component/UiText.kt
@@ -1,4 +1,4 @@
-package com.denihilhamsyah.totphub.totp.presentation
+package com.denihilhamsyah.totphub.totp.presentation.component
 
 import android.content.Context
 import androidx.annotation.StringRes
